Refactor MiscellaneousService to utilize a new get_non_metadata_provider function for improved clarity and maintainability. Updated update_person_and_notify_users method to streamline user notification handling. Ensured compliance with coding standards by maintaining the order of attributes based on their length.
