+++ crates/services/miscellaneous/src/lib.rs
@@ -53,13 +53,12 @@ use dependent_utils::{
     commit_person, create_partial_metadata, deploy_after_handle_media_seen_tasks,
     deploy_background_job, deploy_update_metadata_group_job, deploy_update_metadata_job,
     deploy_update_person_job, first_metadata_image_as_url, get_entity_recently_consumed,
-    get_entity_title_from_id_and_lot, get_google_books_service, get_hardcover_service,
-    get_metadata_provider, get_openlibrary_service, get_tmdb_non_media_service,
-    get_users_and_cte_monitoring_entity, get_users_monitoring_entity,
-    handle_after_media_seen_tasks, is_metadata_finished_by_user, metadata_images_as_urls,
-    post_review, progress_update, refresh_collection_to_entity_association,
-    remove_entity_from_collection, send_notification_for_user, update_metadata_and_notify_users,
-    update_metadata_group, user_metadata_groups_list, user_metadata_list, user_people_list,
+    get_entity_title_from_id_and_lot, get_metadata_provider, get_non_metadata_provider,
+    get_users_monitoring_entity, handle_after_media_seen_tasks, is_metadata_finished_by_user,
+    metadata_images_as_urls, post_review, progress_update, remove_entity_from_collection,
+    send_notification_for_user, update_metadata_and_notify_users, update_metadata_group,
+    update_person_and_notify_users, user_metadata_groups_list, user_metadata_list,
+    user_people_list,
 };
 use enum_meta::Meta;
 use enum_models::{
@@ -74,24 +73,18 @@ use media_models::{
     CreateReviewCommentInput, GenreDetailsInput, GenreListItem, GraphqlCalendarEvent,
     GraphqlMediaAssets, GraphqlMetadataDetails, GraphqlMetadataGroup, GraphqlVideoAsset,
     GroupedCalendarEvent, ImportOrExportItemReviewComment, MarkEntityAsPartialInput,
-    MediaAssociatedPersonStateChanges, MetadataCreator, MetadataCreatorGroupedByRole,
-    MetadataFreeCreator, MetadataImage, MetadataPartialDetails, MetadataVideo, MetadataVideoSource,
-    PersonDetailsGroupedByRole, PersonDetailsItemWithCharacter, PodcastSpecifics,
-    ProgressUpdateInput, ReviewPostedEvent, SeenAnimeExtraInformation, SeenPodcastExtraInformation,
-    SeenShowExtraInformation, ShowSpecifics, UniqueMediaIdentifier, UpdateCustomMetadataInput,
-    UpdateSeenItemInput, UserCalendarEventInput, UserMediaNextEntry,
-    UserMetadataDetailsEpisodeProgress, UserMetadataDetailsShowSeasonProgress,
+    MetadataCreator, MetadataCreatorGroupedByRole, MetadataFreeCreator, MetadataImage,
+    MetadataPartialDetails, MetadataVideo, MetadataVideoSource, PersonDetailsGroupedByRole,
+    PersonDetailsItemWithCharacter, PodcastSpecifics, ProgressUpdateInput, ReviewPostedEvent,
+    SeenAnimeExtraInformation, SeenPodcastExtraInformation, SeenShowExtraInformation,
+    ShowSpecifics, UpdateCustomMetadataInput, UpdateSeenItemInput, UserCalendarEventInput,
+    UserMediaNextEntry, UserMetadataDetailsEpisodeProgress, UserMetadataDetailsShowSeasonProgress,
     UserUpcomingCalendarEventInput,
 };
 use migrations::{
     AliasedCalendarEvent, AliasedMetadata, AliasedMetadataToGenre, AliasedSeen, AliasedUserToEntity,
 };
 use nanoid::nanoid;
-use providers::{
-    anilist::NonMediaAnilistService, audible::AudibleService, igdb::IgdbService,
-    itunes::ITunesService, listennotes::ListennotesService, mal::NonMediaMalService,
-    manga_updates::MangaUpdatesService, vndb::VndbService, youtube_music::YoutubeMusicService,
-};
 use rust_decimal_macros::dec;
 use sea_orm::{
     ActiveModelTrait, ActiveValue, ColumnTrait, ConnectionTrait, DatabaseBackend,
@@ -105,11 +98,9 @@ use sea_query::{
 };
 use serde::{Deserialize, Serialize};
 use supporting_service::SupportingService;
-use traits::{MediaProvider, TraceOk};
+use traits::TraceOk;
 use user_models::DashboardElementLot;
 use uuid::Uuid;
-
-type Provider = Box<(dyn MediaProvider + Send + Sync)>;
 
 pub struct MiscellaneousService(pub Arc<SupportingService>);
 
@@ -1279,7 +1270,7 @@ impl MiscellaneousService {
             return Ok(SearchResults::default());
         }
         let preferences = user_by_id(user_id, &self.0).await?.preferences;
-        let provider = self.get_non_metadata_provider(input.source).await?;
+        let provider = get_non_metadata_provider(input.source, &self.0).await?;
         let results = provider
             .people_search(
                 &query,
@@ -1324,35 +1315,6 @@ impl MiscellaneousService {
         )
         .await?;
         Ok(results)
-    }
-
-    async fn get_non_metadata_provider(&self, source: MediaSource) -> Result<Provider> {
-        let err = || Err(Error::new("This source is not supported".to_owned()));
-        let service: Provider = match source {
-            MediaSource::YoutubeMusic => Box::new(YoutubeMusicService::new().await),
-            MediaSource::Hardcover => Box::new(get_hardcover_service(&self.0.config).await?),
-            MediaSource::Vndb => Box::new(VndbService::new(&self.0.config.visual_novels).await),
-            MediaSource::Openlibrary => Box::new(get_openlibrary_service(&self.0.config).await?),
-            MediaSource::Itunes => {
-                Box::new(ITunesService::new(&self.0.config.podcasts.itunes).await)
-            }
-            MediaSource::GoogleBooks => Box::new(get_google_books_service(&self.0.config).await?),
-            MediaSource::Audible => {
-                Box::new(AudibleService::new(&self.0.config.audio_books.audible).await)
-            }
-            MediaSource::Listennotes => Box::new(ListennotesService::new(self.0.clone()).await),
-            MediaSource::Igdb => Box::new(IgdbService::new(self.0.clone()).await),
-            MediaSource::MangaUpdates => Box::new(
-                MangaUpdatesService::new(&self.0.config.anime_and_manga.manga_updates).await,
-            ),
-            MediaSource::Tmdb => Box::new(get_tmdb_non_media_service(&self.0).await?),
-            MediaSource::Anilist => {
-                Box::new(NonMediaAnilistService::new(&self.0.config.anime_and_manga.anilist).await)
-            }
-            MediaSource::Mal => Box::new(NonMediaMalService::new().await),
-            MediaSource::Custom => return err(),
-        };
-        Ok(service)
     }
 
     pub async fn commit_metadata_group(&self, input: CommitMediaInput) -> Result<StringIdObject> {
@@ -2166,179 +2128,12 @@ impl MiscellaneousService {
         Ok(())
     }
 
-    async fn update_person(
-        &self,
-        person_id: String,
-    ) -> Result<Vec<(String, UserNotificationContent)>> {
-        let person = Person::find_by_id(person_id.clone())
-            .one(&self.0.db)
-            .await?
-            .unwrap();
-        if !person.is_partial.unwrap_or_default() {
-            return Ok(vec![]);
-        }
-        let mut notifications = vec![];
-        let provider = self.get_non_metadata_provider(person.source).await?;
-        let provider_person = provider
-            .person_details(&person.identifier, &person.source_specifics)
-            .await?;
-        ryot_log!(debug, "Updating person for {:?}", person_id);
-        let images = provider_person
-            .images
-            .map(|images| {
-                images
-                    .into_iter()
-                    .map(|i| MetadataImage {
-                        url: StoredUrl::Url(i),
-                    })
-                    .collect()
-            })
-            .filter(|i: &Vec<MetadataImage>| !i.is_empty());
-        let mut current_state_changes = person.clone().state_changes.unwrap_or_default();
-        let mut to_update_person: person::ActiveModel = person.clone().into();
-        to_update_person.images = ActiveValue::Set(images);
-        to_update_person.is_partial = ActiveValue::Set(Some(false));
-        to_update_person.name = ActiveValue::Set(provider_person.name);
-        to_update_person.last_updated_on = ActiveValue::Set(Utc::now());
-        to_update_person.place = ActiveValue::Set(provider_person.place);
-        to_update_person.gender = ActiveValue::Set(provider_person.gender);
-        to_update_person.website = ActiveValue::Set(provider_person.website);
-        to_update_person.source_url = ActiveValue::Set(provider_person.source_url);
-        to_update_person.birth_date = ActiveValue::Set(provider_person.birth_date);
-        to_update_person.death_date = ActiveValue::Set(provider_person.death_date);
-        to_update_person.description = ActiveValue::Set(provider_person.description);
-        to_update_person.alternate_names = ActiveValue::Set(provider_person.alternate_names);
-        for data in provider_person.related_metadata.clone() {
-            let title = data.metadata.title.clone();
-            let pm = create_partial_metadata(data.metadata, &self.0.db).await?;
-            let already_intermediate = MetadataToPerson::find()
-                .filter(metadata_to_person::Column::MetadataId.eq(&pm.id))
-                .filter(metadata_to_person::Column::PersonId.eq(&person_id))
-                .filter(metadata_to_person::Column::Role.eq(&data.role))
-                .one(&self.0.db)
-                .await?;
-            if already_intermediate.is_none() {
-                let intermediate = metadata_to_person::ActiveModel {
-                    role: ActiveValue::Set(data.role.clone()),
-                    metadata_id: ActiveValue::Set(pm.id.clone()),
-                    person_id: ActiveValue::Set(person.id.clone()),
-                    character: ActiveValue::Set(data.character.clone()),
-                    ..Default::default()
-                };
-                intermediate.insert(&self.0.db).await.unwrap();
-            }
-            let search_for = MediaAssociatedPersonStateChanges {
-                role: data.role.clone(),
-                media: UniqueMediaIdentifier {
-                    lot: pm.lot,
-                    source: pm.source,
-                    identifier: pm.identifier.clone(),
-                },
-            };
-            if !current_state_changes
-                .metadata_associated
-                .contains(&search_for)
-            {
-                notifications.push((
-                    format!(
-                        "{} has been associated with {} as {}",
-                        person.name, title, data.role
-                    ),
-                    UserNotificationContent::PersonMetadataAssociated,
-                ));
-                current_state_changes.metadata_associated.insert(search_for);
-            }
-        }
-        for (idx, data) in provider_person.related_metadata_groups.iter().enumerate() {
-            let db_dg = match MetadataGroup::find()
-                .filter(metadata_group::Column::Lot.eq(data.metadata_group.lot))
-                .filter(metadata_group::Column::Source.eq(data.metadata_group.source))
-                .filter(metadata_group::Column::Identifier.eq(&data.metadata_group.identifier))
-                .one(&self.0.db)
-                .await?
-            {
-                Some(m) => m.id,
-                None => {
-                    let m = metadata_group::ActiveModel {
-                        is_partial: ActiveValue::Set(Some(true)),
-                        lot: ActiveValue::Set(data.metadata_group.lot),
-                        source: ActiveValue::Set(data.metadata_group.source),
-                        title: ActiveValue::Set(data.metadata_group.title.clone()),
-                        identifier: ActiveValue::Set(data.metadata_group.identifier.clone()),
-                        images: ActiveValue::Set(
-                            data.metadata_group.images.clone().filter(|i| !i.is_empty()),
-                        ),
-                        ..Default::default()
-                    };
-                    m.insert(&self.0.db).await?.id
-                }
-            };
-            let already_intermediate = MetadataGroupToPerson::find()
-                .filter(metadata_group_to_person::Column::Role.eq(&data.role))
-                .filter(metadata_group_to_person::Column::PersonId.eq(&person_id))
-                .filter(metadata_group_to_person::Column::MetadataGroupId.eq(&db_dg))
-                .one(&self.0.db)
-                .await?;
-            if already_intermediate.is_none() {
-                let intermediate = metadata_group_to_person::ActiveModel {
-                    role: ActiveValue::Set(data.role.clone()),
-                    metadata_group_id: ActiveValue::Set(db_dg),
-                    index: ActiveValue::Set(idx.try_into().unwrap()),
-                    person_id: ActiveValue::Set(person_id.to_owned()),
-                };
-                intermediate.insert(&self.0.db).await?;
-            }
-            let search_for = MediaAssociatedPersonStateChanges {
-                role: data.role.clone(),
-                media: UniqueMediaIdentifier {
-                    lot: data.metadata_group.lot,
-                    source: data.metadata_group.source,
-                    identifier: data.metadata_group.identifier.clone(),
-                },
-            };
-            if !current_state_changes
-                .metadata_groups_associated
-                .contains(&search_for)
-            {
-                notifications.push((
-                    format!(
-                        "{} has been associated with {} as {}",
-                        person.name, data.metadata_group.title, data.role
-                    ),
-                    UserNotificationContent::PersonMetadataGroupAssociated,
-                ));
-                current_state_changes
-                    .metadata_groups_associated
-                    .insert(search_for);
-            }
-        }
-        to_update_person.state_changes = ActiveValue::Set(Some(current_state_changes));
-        to_update_person.update(&self.0.db).await.unwrap();
-        Ok(notifications)
-    }
-
     pub async fn update_metadata_and_notify_users(&self, metadata_id: &String) -> Result<()> {
         update_metadata_and_notify_users(metadata_id, &self.0).await
     }
 
     pub async fn update_person_and_notify_users(&self, person_id: &String) -> Result<()> {
-        let notifications = self.update_person(person_id.clone()).await?;
-        if !notifications.is_empty() {
-            let users_to_notify =
-                get_users_and_cte_monitoring_entity(person_id, EntityLot::Person, &self.0.db)
-                    .await?;
-            for notification in notifications {
-                for (user_id, cte_id) in users_to_notify.iter() {
-                    send_notification_for_user(user_id, &self.0, &notification)
-                        .await
-                        .trace_ok();
-                    refresh_collection_to_entity_association(cte_id, &self.0.db)
-                        .await
-                        .trace_ok();
-                }
-            }
-        }
-        Ok(())
+        update_person_and_notify_users(person_id, &self.0).await
     }
 
     pub async fn update_metadata_group(&self, metadata_group_id: &str) -> Result<()> {
